feat: Implement workspace server with file operations, code execution, and export tools
- Added `file_operations.rs` for handling file read/write, directory listing, and file searching.
- Introduced `mod.rs` to organize workspace-related modules and tools.
- Created `code_tools.rs` for executing Python, TypeScript, and shell commands.
- Developed `export_tools.rs` for exporting files and ZIP packages from the workspace.
- Implemented `file_tools.rs` for reading, writing, and manipulating files.
- Added `ui_resources.rs` for generating UI responses and HTML for exports.
- Created `utils.rs` for common utility functions and constants related to workspace operations.
- Enhanced the overall structure and organization of the workspace module for better maintainability.
