+++ docs/history/refactoring_20250904_1000.md
@@ -1,0 +1,204 @@
+# Remote DOM 기반 MCP-UI 리팩터링 계획
+
+## 작업의 목적
+
+현재 HTML UIResource(text/html) 기반으로 구현된 MCP-UI 컴포넌트를 Remote DOM(application/vnd.mcp-ui.remote-dom) 방식으로 변경하여, 호스트 앱의 룩앤필과 일치하는 네이티브 스타일 UI 제공과 더 나은 사용자 경험을 달성한다.
+
+## 현재의 상태 / 문제점
+
+### 현재 구현 상태
+
+- **Frontend**: `@mcp-ui/client`와 `UIResourceRenderer`를 통해 HTML UIResource를 iframe으로 렌더링
+- **Backend**: `WorkspaceServer`에서 `text/html` mimeType으로 HTML 문자열 직접 생성
+- **지원되는 컨텐츠 타입**: HTML, External URL만 지원 (Remote DOM 미지원)
+
+### 주요 문제점
+
+1. **룩앤필 이질감**: iframe으로 렌더링되는 HTML UI가 호스트 앱의 Tailwind/shadcn 스타일과 다름
+2. **제한된 상호작용**: iframe 샌드박스 환경으로 인한 호스트와의 제한적 통신
+3. **스타일 일관성 부족**: 각 HTML UIResource마다 독립적인 CSS로 인한 디자인 불일치
+4. **테마 지원 부족**: 호스트의 다크/라이트 테마가 iframe 내부에 반영되지 않음
+
+### 구체적 이슈
+
+- `WorkspaceServer.create_html_export_ui()`: 하드코딩된 CSS 스타일로 호스트와 이질적
+- `UIResourceRenderer`: Remote DOM 타입 처리 로직 부재
+- 패키지 의존성: `@remote-dom/core` 미설치 상태
+
+## 추가 분석 과제
+
+1. **@remote-dom/core 패키지 분석**
+   - 설치 후 API 구조 및 사용 패턴 파악
+   - React 컴포넌트 생성 및 렌더링 방식 이해
+   - Shopify의 remote-dom 프레임워크 동작 원리 학습
+
+2. **MCP-UI Remote DOM 표준 조사**
+   - `application/vnd.mcp-ui.remote-dom+javascript; framework=react` mimeType 스펙
+   - 클라이언트-서버 간 컴포넌트 정의 및 이벤트 통신 프로토콜
+   - 기존 HTML 기반 postMessage와의 차이점 분석
+
+3. **컴포넌트 라이브러리 설계**
+   - 호스트의 shadcn/ui 컴포넌트를 Remote DOM에서 사용하는 방법
+   - 테마 시스템 연동 방안 (CSS Variables, Tailwind Classes)
+   - 재사용 가능한 기본 컴포넌트 세트 정의
+
+## 변경 이후의 상태 / 해결 판정 기준
+
+### 목표 상태
+
+1. **일관된 룩앤필**: Remote DOM 컴포넌트가 호스트의 Tailwind/shadcn 스타일과 완전히 일치
+2. **네이티브 상호작용**: iframe 없이 직접 React 컴포넌트로 렌더링되어 자연스러운 UX
+3. **테마 자동 적용**: 호스트의 다크/라이트 테마가 자동으로 MCP-UI 컴포넌트에 반영
+4. **확장 가능한 구조**: 새로운 컴포넌트 타입을 쉽게 추가할 수 있는 유연한 아키텍처
+
+### 해결 판정 기준
+
+- [ ] `WorkspaceServer`에서 `application/vnd.mcp-ui.remote-dom` mimeType 지원
+- [ ] `UIResourceRenderer`에서 Remote DOM 컴포넌트 정상 렌더링
+- [ ] Export UI가 호스트 앱과 동일한 스타일 적용
+- [ ] 다크/라이트 테마 전환 시 MCP-UI 컴포넌트도 함께 변경
+- [ ] 기존 HTML UIResource 대비 동일한 기능 제공 (다운로드, 파일 목록 등)
+
+## 수정이 필요한 코드 및 수정부분의 코드 스니핏
+
+### 1. 패키지 의존성 추가
+
+```json
+// package.json
+{
+  "dependencies": {
+    "@remote-dom/core": "^2.0.0"
+  }
+}
+```
+
+### 2. Frontend: UIResourceRenderer 확장
+
+```typescript
+// src/components/ui/UIResourceRenderer.tsx
+import { RemoteDOMResourceRenderer } from '@mcp-ui/client';
+
+const UIResourceRenderer: React.FC<UIResourceRendererProps> = ({
+  resource,
+  onUIAction,
+}) => {
+  // Remote DOM 지원 추가
+  return (
+    <ExternalUIResourceRenderer
+      resource={mcpUIResource}
+      onUIAction={handleUIAction}
+      supportedContentTypes={['rawHtml', 'externalUrl', 'remoteDom']} // remoteDom 추가
+      remoteDomProps={{
+        library: customComponentLibrary, // shadcn/ui 기반 컴포넌트 라이브러리
+      }}
+    />
+  );
+};
+```
+
+### 3. Backend: WorkspaceServer Remote DOM 지원
+
+```rust
+// src-tauri/src/mcp/builtin/workspace.rs
+impl WorkspaceServer {
+    fn create_export_remote_dom_ui(
+        &self,
+        title: &str,
+        files: &[String],
+        export_type: &str,
+        download_path: &str,
+    ) -> String {
+        // Remote DOM 스크립트 생성
+        format!(
+            r#"
+            const container = document.createElement('ui-card');
+            container.setAttribute('class', 'p-6 space-y-4');
+
+            const header = document.createElement('ui-card-header');
+            const title = document.createElement('ui-card-title');
+            title.textContent = '{}';
+            header.appendChild(title);
+
+            const content = document.createElement('ui-card-content');
+            // ... Remote DOM 컴포넌트 구성
+
+            container.appendChild(header);
+            container.appendChild(content);
+            root.appendChild(container);
+            "#,
+            title
+        )
+    }
+
+    fn create_export_ui_resource(&self, ...) -> Value {
+        json!({
+            "type": "resource",
+            "resource": {
+                "uri": format!("ui://export/remote/{}", request_id),
+                "mimeType": "application/vnd.mcp-ui.remote-dom+javascript; framework=react",
+                "text": self.create_export_remote_dom_ui(...), // Remote DOM 스크립트
+                // ...
+            }
+        })
+    }
+}
+```
+
+### 4. 컴포넌트 라이브러리 생성
+
+```typescript
+// src/lib/mcp-ui/component-library.ts
+import { ComponentLibrary } from '@mcp-ui/client';
+import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card';
+import { Button } from '@/components/ui/button';
+
+export const customComponentLibrary: ComponentLibrary = {
+  'ui-card': Card,
+  'ui-card-header': CardHeader,
+  'ui-card-title': CardTitle,
+  'ui-card-content': CardContent,
+  'ui-button': Button,
+  // ... 추가 컴포넌트 매핑
+};
+```
+
+## 재사용 가능한 연관 코드
+
+### 관련 파일 및 주요 기능
+
+1. **`src/components/ui/UIResourceRenderer.tsx`**
+   - 기능: MCP-UI 리소스 렌더링 (현재 HTML/URL만 지원)
+   - 인터페이스: `UIResourceRendererProps`, `UIResource`
+   - 재사용: Remote DOM 지원 확장, 컴포넌트 라이브러리 통합
+
+2. **`src-tauri/src/mcp/builtin/workspace.rs`**
+   - 기능: 파일 시스템 작업 및 HTML UIResource 생성
+   - 주요 메서드: `create_export_ui_resource()`, `create_html_export_ui()`
+   - 재사용: Remote DOM 스크립트 생성 로직 추가
+
+3. **`src/models/chat.ts`**
+   - 기능: UIResource 타입 정의
+   - 인터페이스: `UIResource` (mimeType 확장 필요)
+   - 재사용: Remote DOM mimeType 지원 추가
+
+4. **`src/lib/mcp-types.ts`**
+   - 기능: MCP 표준 타입 정의
+   - 인터페이스: `MCPResourceContent`, `UIResource`
+   - 재사용: Remote DOM 관련 타입 확장
+
+5. **`src/components/ui/` (shadcn/ui 컴포넌트들)**
+   - 기능: 재사용 가능한 UI 컴포넌트 라이브러리
+   - 주요 컴포넌트: Card, Button, Badge, Input 등
+   - 재사용: Remote DOM 컴포넌트 라이브러리의 기반
+
+### 기존 MCP-UI 통합 패턴
+
+- `@mcp-ui/client` 패키지 활용 방식
+- UIAction 이벤트 처리 패턴 (`onUIAction` 콜백)
+- postMessage 기반 호스트-컴포넌트 통신 구조
+
+### 스타일링 시스템
+
+- Tailwind CSS 기반 유틸리티 클래스
+- CSS Variables를 통한 테마 시스템
+- shadcn/ui의 디자인 토큰 및 컴포넌트 구조
